--- frontend/src/pages/Home/index.jsx
@@ -1,1 +1,0 @@
-export {Home} from "./Home";
